# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Banter Bus",
    "description": "The API definition for the Banter Bus server.",
    "version": "1.0.0"
  },
  "paths": {
    "/game": {
      "get": {
        "tags": [
          "game"
        ],
        "summary": "Get all game types.",
        "operationId": "GetAllGameType",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "game"
        ],
        "summary": "Create a new game type.",
        "operationId": "CreateGameType",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "rules_url": {
                    "type": "string"
                  }
                },
                "required": [
                  "name",
                  "rules_url"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsGame"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/game/{name}": {
      "get": {
        "tags": [
          "game"
        ],
        "summary": "Get a game types.",
        "operationId": "GetGameType",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsGame"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "game"
        ],
        "summary": "Delete a game types.",
        "operationId": "RemoveGameType",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/game/{name}/disable": {
      "put": {
        "tags": [
          "game"
        ],
        "summary": "Disables a game type.",
        "operationId": "DisableGameType",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/game/{name}/enable": {
      "put": {
        "tags": [
          "game"
        ],
        "summary": "Enables a game type.",
        "operationId": "EnableGameType",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsGame"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/game/{name}/question": {
      "post": {
        "tags": [
          "question"
        ],
        "summary": "Add a new question to a game type.",
        "operationId": "AddQuestion",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string"
                  },
                  "round": {
                    "type": "string"
                  }
                },
                "required": [
                  "content",
                  "round"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ModelsGame": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "rules_url": {
            "type": "string"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "game",
      "description": "Related to managing the game types."
    },
    {
      "name": "question",
      "description": "Related to managing the questions to a game type."
    }
  ]
}